+++ examples/factories/project-factory/main.tf
@@ -28,7 +28,7 @@ locals {
   _service_accounts_iam = {
     for r in local._service_accounts_iam_bindings : r => [
       for k, v in var.service_accounts :
-      "serviceAccount:${k}@${local._project_id}.iam.gserviceaccount.com"
+      module.service-accounts[k].iam_email
       if try(index(v, r), null) != null
     ]
   }