+++ fast/stages/01-resman/main.tf
@@ -16,6 +16,13 @@
 
 locals {
   # convenience flags that express where billing account resides
+  automation_resman_sa = try(
+    [format(
+      "serviceAccount:%s",
+      data.google_client_openid_userinfo.provider_identity.0.email
+    )],
+    []
+  )
   billing_ext     = var.billing_account.organization_id == null
   billing_org     = var.billing_account.organization_id == var.organization.id
   billing_org_ext = !local.billing_ext && !local.billing_org
@@ -24,11 +31,33 @@ locals {
     if(
       v != null
       &&
-      contains(keys(local.identity_providers), try(v.identity_provider, ""))
+      (
+        try(v.type, null) == "sourcerepo"
+        ||
+        contains(
+          keys(local.identity_providers),
+          coalesce(try(v.identity_provider, null), ":")
+        )
+      )
       &&
       fileexists("${path.module}/templates/workflow-${try(v.type, "")}.yaml")
     )
   }
+  cicd_workflow_var_files = {
+    stage_2 = [
+      "00-bootstrap.auto.tfvars.json",
+      "01-resman.auto.tfvars.json",
+      "globals.auto.tfvars.json"
+    ]
+    stage_3 = [
+      "00-bootstrap.auto.tfvars.json",
+      "01-resman.auto.tfvars.json",
+      "globals.auto.tfvars.json",
+      "02-networking.auto.tfvars.json",
+      "02-security.auto.tfvars.json"
+    ]
+  }
+
   custom_roles = coalesce(var.custom_roles, {})
   groups = {
     for k, v in var.groups :
@@ -41,4 +70,8 @@ locals {
   identity_providers = coalesce(
     try(var.automation.federated_identity_providers, null), {}
   )
+}
+
+data "google_client_openid_userinfo" "provider_identity" {
+  count = length(local.cicd_repositories) > 0 ? 1 : 0
 }