+++ fast/stages/01-resman/outputs.tf
@@ -15,60 +15,48 @@
  */
 
 locals {
-  _cicd_tf_var_files = {
-    stage_2 = [
-      "00-bootstrap.auto.tfvars.json",
-      "01-resman.auto.tfvars.json",
-      "globals.auto.tfvars.json"
-    ]
-    stage_3 = [
-      "00-bootstrap.auto.tfvars.json",
-      "01-resman.auto.tfvars.json",
-      "globals.auto.tfvars.json",
-      "02-networking.auto.tfvars.json",
-      "02-security.auto.tfvars.json"
-    ]
-  }
   _tpl_providers = "${path.module}/templates/providers.tf.tpl"
   cicd_workflow_attrs = {
     data_platform_dev = {
       service_account   = try(module.branch-dp-dev-sa-cicd.0.email, null)
       tf_providers_file = "03-data-platform-dev-providers.tf"
-      tf_var_files      = local._cicd_tf_var_files.stage_3
+      tf_var_files      = local.cicd_workflow_var_files.stage_3
     }
     data_platform_prod = {
       service_account   = try(module.branch-dp-prod-sa-cicd.0.email, null)
       tf_providers_file = "03-data-platform-prod-providers.tf"
-      tf_var_files      = local._cicd_tf_var_files.stage_3
+      tf_var_files      = local.cicd_workflow_var_files.stage_3
     }
     networking = {
       service_account   = try(module.branch-network-sa-cicd.0.email, null)
       tf_providers_file = "02-networking-providers.tf"
-      tf_var_files      = local._cicd_tf_var_files.stage_2
+      tf_var_files      = local.cicd_workflow_var_files.stage_2
     }
     project_factory_dev = {
-      service_account   = try(module.branch-pf-dev-sa-cicd.0.email, null)
+      service_account   = try(module.branch-teams-dev-pf-sa-cicd.0.email, null)
       tf_providers_file = "03-project-factory-dev-providers.tf"
-      tf_var_files      = local._cicd_tf_var_files.stage_3
+      tf_var_files      = local.cicd_workflow_var_files.stage_3
     }
     project_factory_prod = {
-      service_account   = try(module.branch-pf-prod-sa-cicd.0.email, null)
+      service_account   = try(module.branch-teams-prod-pf-sa-cicd.0.email, null)
       tf_providers_file = "03-project-factory-prod-providers.tf"
-      tf_var_files      = local._cicd_tf_var_files.stage_3
+      tf_var_files      = local.cicd_workflow_var_files.stage_3
     }
     security = {
       service_account   = try(module.branch-security-sa-cicd.0.email, null)
       tf_providers_file = "02-security-providers.tf"
-      tf_var_files      = local._cicd_tf_var_files.stage_2
+      tf_var_files      = local.cicd_workflow_var_files.stage_2
     }
   }
   cicd_workflows = {
     for k, v in local.cicd_repositories : k => templatefile(
       "${path.module}/templates/workflow-${v.type}.yaml",
       merge(local.cicd_workflow_attrs[k], {
-        identity_provider = local.identity_providers[v.identity_provider].name
-        outputs_bucket    = var.automation.outputs_bucket
-        stage_name        = k
+        identity_provider = try(
+          local.identity_providers[v.identity_provider].name, null
+        )
+        outputs_bucket = var.automation.outputs_bucket
+        stage_name     = k
       })
     )
   }
@@ -158,9 +146,11 @@ output "cicd_repositories" {
   description = "WIF configuration for CI/CD repositories."
   value = {
     for k, v in local.cicd_repositories : k => {
-      branch          = v.branch
-      name            = v.name
-      provider        = local.identity_providers[v.identity_provider].name
+      branch = v.branch
+      name   = v.name
+      provider = try(
+        local.identity_providers[v.identity_provider].name, null
+      )
       service_account = local.cicd_workflow_attrs[k].service_account
     } if v != null
   }