+++ modules/iam-service-account/outputs.tf
@@ -30,6 +30,14 @@ output "iam_email" {
   ]
 }
 
+output "id" {
+  description = "Service account id."
+  value       = local.service_account.id
+  depends_on = [
+    local.service_account
+  ]
+}
+
 output "key" {
   description = "Service account key."
   sensitive   = true
@@ -37,7 +45,7 @@ output "key" {
 }
 
 output "name" {
-  description = "Service account id."
+  description = "Service account name."
   value       = local.service_account.name
 }